Bump cluster-services chart version.
Forgot to do this in #348.

Since we install this chart via `helm install` (via Terraform), we need
the version number change to trigger chart-releaser to build the new
tarball of the chart. If we don't bump the version, the cluster turnup
automation will break because it'll be stuck on an old version of the
chart.