Add Helm values for ArgoCD apps in the staging environment.
Non-frontend apps (except Signon) are intentionally excluded at this
time.

This was generated from values-integration.yaml by deleting the unwanted
apps and performing a global replace of `s/integration/staging`.

Updating the config items to the actual values for staging will come in
a separate commit, so that it's easier to see what's going on.