Stop relying on aws-cli in the app container in upload-assets.
Instead, run the app image as an initContainer to copy the assets files
to an emptyDir ephemeral volume, then use the aws-cli image (the Bitnami
one, hosted on Amazon ECR Public Gallery) to upload the files to S3.

Tested: created a test ArgoCD app in the integration cluster by copying
the frontend one, changing its name and pointing its targetRevision at
the PR branch. Assets upload works (very quickly) and puts the files in
the right place in S3.