Use the ECR-hosted Bitnami aws-cli image for static-error-page-upload.
This brings it in line with assets-upload-job.

The security/support situation for the Bitnami one is at least as good
as the Amazon one, plus it's hosted on ECR. We don't want to pull images
from DockerHub.

Perplexingly, Amazon doesn't seem to host its own aws-cli images on its
own public ECR mirror, but it does host the Bitnami one.

While we're there, sync up some of the settings with assets-upload-job:
- revert to the default backoffLimit of 6 (relying on Argo to retry if
  the service takes a while to come up seems a bit
  implementation-specific)
- don't pin the aws-cli minor version (the syntax isn't going to change
  and we will get burned if we don't keep up to date)
- revert to the default hook-delete-policy (delete before creating a new
  one, otherwise you have to delete a failed job manually and cancel the
  Argo sync in order to recovery after a failed job).