Remove cluster-secrets chart from ArgoCD.
It's risky and problematic to deploy this from both Terraform and
ArgoCD. It leads to a race that breaks cluster turnup in most cases. We
could work around that (e.g. with depends_on and ignore_changes in
Terraform) but to do so would be to add complexity and brittleness just
to preserve something which is fairly self-evidently undesirable:
having two different deployment tools managing the same objects is unwise.