Remove htpasswd volumemounts in staging.
Unlike integration, staging uses an IP address allowlist instead of
Basic auth. (Presumably so that it's a more realistic model of
production with regard to things like HTTP caching.) We want to keep
this the same as the EC2 stack at least for the time being.

Basic auth is already conditional on the environment in the nginx
configs, but we were trying to mount the secret even in environments
where it doesn't exist.