Restrict staging www-origin ingress to GDS and Fastly netblocks.
Trusting the network is not good practice, but in this case improving on
this pre-existing arrangement is nontrivial and not in scope for the
Replatforming project, so we're just keeping the status quo for now.
In other words, we're replicating the same access control that we
already have on the current, EC2 staging environment.

Ingresses for publisher apps (including Signon) are excluded from this
control because that's how it is in EC2 staging (presumably because
Signon is the point of access control for the non-public-facing stuff).
We'll need to apply this allow-list to assets-origin when we write the
config for that. (assets-origin doesn't exist yet in our k8s rig.)