Update licencefinder to latest trunk build.
The deployment webhook hasn't run recently (not sure why, might just be
that it was broken for a while and the last successful build was before
we finished the deployment pipeline). We need a more recent image
running in staging in order for the app to work, hence this one-off
manual update.