Update image tag script to not fail when tag already set
This makes the update image tag script exit successfully when the image
tag is already set to the desired value. Previously the script would
exit with an error as there was nothing to commit. This causes the
pipeline to look as if it has failed and argo workflows will attempt to
retry the job.