Avoid creating backend ExternalSecrets in prod/staging for now.
We'll remove these conditionals when we're ready to deploy the rest of
the stack (i.e. backends and APIs) to staging/prod.

For now, this stops the govuk-apps-conf Argo app from crying wolf
because some of the Secrets Manager secrets don't exist (and we don't
want to set those secrets now just to silence it, because then we'll
have to keep checking for drift between EC2 prod/staging and our
not-ready-yet prod/staging).