Make it easier to debug the error-page-upload job.
Allow enabling the job by setting a Helm value. This allows running the
job when deploying Static under a different name.

Also pass the service name for Static into the script so that it uses
the right one when testing (again, when the chart is deployed under a
different name).

Enable tracing in the script so that it's clearer from the logs as to
what's running when.

Add a description to the Job object.