Reuse the same connection when scraping error pages from Static.
It seems `curl -f` bases its exit code on the status of the last
request, even with `--fail-early`, hence the additional check to make
sure we really downloaded all the files. I think the additional check is
worth having anyway, especially since we've already had one bug in that
area.

See the [curl manpage] for explanation of how the substitution in `-o`
works.

[curl manpage]: https://curl.se/docs/manpage.html

Tested: checked the failure and success cases manually; output looks
good and it's easy to see what went wrong when it fails.