Name containers within pods more consistently.
There's no need to include the name of the app/deployment in the
container name, since it's already in the pod name. On the other hand,
it's super annoying having to keep typing these different, long names
all the time when debugging.

The container names now make sense in the context of the pod (as they
should) and no longer contain any (unhelpful) redundant information.