Use the shared assets filer for whitehall-admin asset uploads.
Whitehall (whitehall-admin) assumes that it's writing uploaded files to
a filesystem that is shared between all Whitehall processes (i.e. all
whitehall-admin pods, including the worker deployment).

We might as well use the same EFS filer as Asset Manager, since this is
essentially the same functionality and purpose. Asset Manager doesn't
need to communicate with Whitehall via the filer, but there's also no
risk of naming conflict.

This is for CMS assets i.e. article images and attachments, not app
CSS/JS etc.