Use same UID/GID for whitehall-admin as the old EC2 stack.
This is the most straightforward way to ensure that whitehall-admin on
k8s can coexist with whitehall-admin on EC2, i.e. that each can read
files that were uploaded by the other.

Some minor refactoring in order to specify runAsUser/runAsGroup, because
so far we've been lumping all the container securityContext stuff behind
a strangely-named boolean `.Values.securityContext.appContainer`.