Prevent manual deploy disabling automatic deploy
This changes the manual deploy behaviour to prevent it from disabling
automatic deployments. This behaviour is consistent with the old
infrastructure so should be less confusing developers. Automatically
disabling automatic deploys was more risky as it depends on developers
remember to switch it back on, which has already been seen to be an
issue.