Remove Signon Resources
This was used to manage signon resources declaritively and synchornise
secrets with Kubernetes. However the implementation is incomplete and
cannot handle Staging being sychronised with Prod everyday. We currently
only need to sync secrets with K8s, which we can do with the External
Secrets and the webhook provider, as we're not concerned with spining
entirely new cluster yet. Therefore we don't want to manage or maintain
the complexity of moving Signon to a declaritive state as of yet. The
API for signon will also be modified, so switching this off to prevent
any issue with breaking changes to the API.