Pass repoName Helm value down to app charts.
This is useful for constructing names of things which are a function of
the app's repo as opposed to the Helm release name.

For example: Sentry DSNs (API keys) are unique per app repo but are not
unique per app. `whitehall-admin`, `whitehall-frontend` and
`draft-whitehall-frontend` are all the same Sentry project and so they
have the same Sentry DSN.