Redirect straight to https for signon during login.
Internal requests (from publisher apps to signon within the cluster)
use this same endpoint URL. Presumably the client library was following
the http->https redirect, otherwise it wouldn't have been working.

Ideally we'd have the internal requests go to http://signon (which the
local resolver will expand to signon.apps.svc.cluster.local.) but we'd
need to modify Plek to support that, which takes time because it
involves a gem version increment for every app.