Allow substituting Helm values inside the extraEnv value.
This seems the least evil way to allow for overriding
`PLEK_SERVICE_*_URI` for the draft apps.

This is a two-edged, poison-tipped blade. Use sparingly and with care.

Tested: inspected template output before and after with e.g.

    helm template ../govuk-rails-app --name-template=draft-content-store --values <(
      helm template . --values values-integration.yaml |
        yq e '. |select(.metadata.name == "draft-content-store").spec.source.helm.values'
    ) | yq e '.|select(.kind == "Deployment")'

No changes to existing output.

Background and rationale:

The draft apps need `PLEK_HOSTNAME_PREFIX=draft-` so that they default
to using the draft flavour of all of their backends. The reason we need
to say "assume everything starts with `draft-`, except for these few
backends" rather than specifying the "draft-" URL for each of the draft
backends explicitly is that we don't want draft apps to talk to
non-draft backends even in the event of a misconfiguration.

There are a few apps that don't have a "draft" flavour: signon,
feedback, info-frontend. In order to override these back to "non-draft",
we'd either need to do so globally (which runs into the problem that
`PLEK_SERVICE_*_URI` overrides both the internal and the external URI to
the same value) or we need to be able to substitute the domain name
suffixes inside extraEnv where we're constructing the URLs for the
`PLEK_SERVICE_*_URI` overrides.