Make Signon app names + API usernames match prod/staging.
Some of these are inconsistent between integration and staging/prod in
the Signon database. Let's use the production names (which are the same
as staging because the database is synced nightly) so that we don't have
to mess with the live configuration in production.

I've updated the names in integration Signon to match the staging/prod
ones here, so this should fix the `Could not find apps for: Content
Publisher, Search admin, Specialist Publisher` and `Could not find api
users for: emailalertfrontend@alphagov.co.uk` errors in the
sync-app-secrets and sync-token-secrets jobs respectively.