Trigger deploy promotion in Production
This changes the post-sync workflow to trigger the deploy image workflow
in Production, instead of running the update image tag workflow in the
current environment. This is because the deploy image workflow in
Production is also configured to retag images with a `deployed-to-<env>`
tag. Currently only the production cluster has permissions to retag
images in AWS ECR (which is also in the Production account). Due to AWS
IAM permissions being to granular ("ecr:PutImage"), permission to retag
an image, also allows you to push a binary, which we didn't want to
grant to our non-production environments.