Run publishing-api AMQP "heartbeat" cron less often.
We can probably get rid of this job without too much trouble - it
seems to be just a workaround for some iffy Icinga monitoring - but for
now let's just set it to run 1/4 as often so it doesn't hose the ArgoCD
UI when it fails.

The threshold for the questionable Icinga monitoring is 5 minutes, so
running this every 4 minutes should be perfectly adequate.