Remove logic to update image tag in set_automatic_deploy_enabled
This removes the ability to change the image tag in the script used to
update the flag for enabling automatic deploys. This script should only
be concerned with one thing. We already have a script and workflow to
update the image tag. This prevents us having multiple places/ways to
update image tags that we have to maintain.