Remove unneeded GOVUK_APP_NAME and GOVUK_APP_ROOT env vars.
GOVUK_APP_NAME is set in the container images (via Dockerfile) so we
don't need to set it here. It stays the same for any given app,
regardless of whether it's the draft or non-draft deployment (and
similarly for whitehall frontend/backend deployments - they're both
`GOVUK_APP_NAME=whitehall`).

GOVUK_APP_ROOT is only supposed to be for the old deployment automation
in Puppet/EC2. It's been (mis)used in a couple of places in app code,
but only in Rake tasks. The Rake tasks in question need some fixing up
anyway, so we can safely get rid of GOVUK_APP_ROOT from our config here.

https://sourcegraph.com/search?q=repo:alphagov+GOVUK_APP_ROOT