Add missing volumeMounts for /tmp.
Turns out we'd been writing to /tmp inside the image filesystem and
getting away with it until recently because we hadn't set
readOnlyRootFilesystem.

This should fix most (maybe all?) of the fallout from #711 / d337216.

Also clean up the indentation on all the templates that we're touching.
This should help to avoid a repeat of #702. (We need a linter for this,
but that could be a bit tricky because we're dealing with YAML generated
by Go templates, rather than just YAML.)