Add IngressGroup for assets domain
This adds ALB ingress annotations for Asset Manager, Whitehall Frontend and Static to create a IngressGroup so the AWS Load Balancer Controller uses a shared ALB. This allows request on the assets domain to routed to the correct application.

This also adds the ability to specify a host in the ingress rules.