Allow requests from assets domain in assets ALB rules
This adds a condition the ALB rules for asset-origin to allow requests
from the public facing asset domain. This is because request are being
forwarded from Fastly, where the Host header is set as the original
asset domain (not the domain of the alb).  This is done by adding an
annotation to the ingress resources that is used to AWS ALB ingress
controller.

The annotation are put through the template function as the key and
value require a Helm value to be rendered.