Run the Signon token sync cron nightly instead of hourly.
I couldn't think of much of an advantage to running it hourly now that
the configuration is fairly stable. When it comes to rotating tokens,
we'll still need a transition period where old and new token are both
valid (which might as well be a day if it's an hour), and it'd be
straightforward for any rotation automation to just start a sync job on
demand.