Move argo workflows back to apps namespace
This is because the post sync workflow needs to trigger the smoke test
workflow (which is in the apps namespace). Workflows need to be in the
same namespace to trigger another workflow. Smoke tests can't trivially
be moved to cluster-services as it relies on other dependencies in the
apps namespace.