Specify hostname in ingress rules instead of tls
This updates the ingress configuration to add hostnames to the rules of
the ingress resource, instead of using the tls configuration to define
which certificates need to be added. This allows the ALB Ingress
Controller to automatically add the hostnames as host-header allow rule
- which we need to allow traffic from origin.

The ALB Ingress Controller uses either the hostname in the rules or the
tls config to find certs to attach, so both configurations have the same
behaviour.

This also adds some logic to generate a list of hostnames to use in the
external-dns annotation. However, we might be able to remove the
annotation anyway as the External DNS can create DNS records based on
the hostnames in the rules.