Fix bad resource names when installed as dependency.
Use the `fullname` template from the default _helpers.tpl instead of
referring directly to `.Release.Name`.

Using `.Release.Name` only works as intended when the chart is being
installed at the top level. When a chart is installed as a dependency of
another chart, `.Release.Name` is the name of the parent chart.

This is the problem that Helm's default templates like `fullname` are
intended to solve.

We also have to avoid using `.Release.Name` as a default for things like
the Rails assets directory name. While we're there, omit the redundant
app name prefix from the container names in the upload-assets job.

Tested: no diffs in `helm template` output except for simplifying the
container names in the upload-assets job.