Set max request body size for file uploads.
This matches the 500 MB limit that we currently set for these via
govuk-puppet. It fixes things like image uploads on content-publisher.

nginx's default is 1M, which should be plenty, so keep that for
everything else.

In govuk-puppet this is set to 20M for all the other apps (i.e. apps
that don't handle file uploads), but I intentionally haven't set that
here because there should be no legit reason for those apps to receive
such huge request bodies; it's just a (small but unnecessary)
resource-exhaustion attack risk. I couldn't find any justification in
the govuk-puppet commit history as to why it was set so large.

Similarly, I couldn't find any explanation as to why search-api had this
set to 500 MB so I'm leaving that at the default 1M too.

Tested: inspected `helm template` output for an app with/without an override of
the new value:

    helm template content-publisher ../generic-govuk-app --values <(helm template . --values values-integration.yaml |yq e '.|select(.metadata.name=="content-publisher").spec.source.helm.values')

    helm template frontend ../generic-govuk-app --values <(helm template . --values values-integration.yaml |yq e '.|select(.metadata.name=="frontend").spec.source.helm.values')