Remove some no-longer-needed config for Whitehall.
The generic-govuk-app chart now defaults to using the repo name for the
S3 Rails assets paths, so we no longer need to override it.

We also don't need to set `NODE_ENV` in the environment, since Node.js
isn't actually used at runtime and is no longer in the Whitehall
container image.