No TLS for Bouncer.
Now that it's no longer inheriting our default set of Ingress
annotations, we can configure Bouncer's ALB as needed.

TLS for transitioned sites terminates at Fastly (for the ones that
have TLS at all!). We never built the integration with Let's Encrypt or
whatever to be able to serve TLS at origin for these sites.