Fix usage of externalDomainSuffix
This replaces reference to the publishingDomainSuffix in the ALB rules,
when actually we mean the externalDomainSuffix. They just happened to be
the same value in integration and staging. Now the config is consistent
across all three environments.

This also updates GOVUK_WEBSITE_ROOT to reference the
testExternalDomainSuffix, as are not serving traffic from the non-test
domains.