Remove dashboard refresh intervals < sampling interval.
Having dashboards defaulting to auto-refresh every 10s when the sampling
interval is 60s isn't the most helpful. Also helps keep the resource
usage under control on the backend.

Also get rid of the shortcut for 5m time window, since that's not going
to be used very frequently when all the metrics are at 1m resolution.