Make app.kubernetes.io/name consistent with app label.
We need these labels to be consistent so that we can use them in
Prometheus.

The [k8s docs] are imprecise about the intended semantics of the
well-known labels, so we have to do some interpretation of our own here.
For our existing `app` label we've always considered the web (`app`)
deployment and `worker` deployment as distinct, because the typical use
case is to select pods belonging to one or other deployment. So let's
stick with that since it seems to be the more pragmatic/useful choice.

If we find we need to select objects across a whole instance of a chart,
then we could still use `part-of` for that. I haven't added that here
though because we don't have a use case for it at the moment.

[k8s docs]: https://kubernetes.io/docs/concepts/overview/working-with-objects/common-labels/#labels