Remove namespace prefix from Prometheus job label.
We already have a separate label for the namespace, so having the job
labels prefixed with the namespace is unnecessary and complicating.

kube-prometheus-operator is prepending the namespace to the job label
because we're not telling it which selector label to use for this. The
default is designed to be "safe" at all costs, but this denormalisation
(so to speak) is problematic because it complicates queries for
dashboards and the like.

Also make the metadata consistent for the PodMonitor objects themselves
and make Asset Manager heed the `monitoring.enabled` Helm value.