Enable memcache for local-links-manager and maslow.
These two apps use memcached in the EC2 environments, but we'd missed
this when configuring them in Kubernetes.

Enable memcache so that we don't have performance regressions for these
apps.

Neither stores a lot of data in memcache or does a lot of througput,
and as of https://github.com/alphagov/maslow/pull/953 they both
namespace their keys appropriately. So it's fine for them to use the
shared memcache.

I suppose this makes the name "frontend-memcached" less than a perfect
fit, but I think it's still a reasonable name because these two apps are
still frontends in all but GOV.UK's unusual usage of the term. (They're
UIs and they're the first thing in the request path after the load
balancer, so they're frontends.)