Make LBs recognise the publishing.service hostnames.
When we come to the go-live switchover for the publisher apps, we're
going to need the load balancers to recognise both the govuk.digital
(testing) hostnames and the "real" publishing.service.gov.uk hostnames
which end-users will use.

The hostnames on the Ingress rules (as opposed to the
`alb.ingress.kubernetes.io/load-balancer-attributes` annotation) need to
be only those hostnames which we want external-dns to create for us.