Set deletionPolicy=Delete on ExternalSecrets.
Set deletionPolicy=Delete on ExternalSecrets, so that when Argo CD
deletes an ExternalSecret and re-creates it with a different uid, the
controller deletes Secret that was owned by the old ExternalSecret and
is then able to create the new Secret without finding a conflict of
ownership.

This hopefully fixes the problem where ExternalSecrets often shows as
"degraded" in Argo CD when replacing an Application resource.

All the changes except the ones to values.yaml were generated with:

    git grep -l ExternalSecret external-secrets asset-manager generic-govuk-app smokey |xargs gsed -i '/  target:/a \ \ \ \ deletionPolicy: {{ .Values.externalSecrets.deletionPolicy }}'

    git grep -l ExternalSecret asset-manager generic-govuk-app smokey |xargs gsed -i 's/refreshInterval: 1h/refreshInterval: {{ .Values.externalSecrets.refreshInterval }}/'