Allow Argo CD to create ClusterExternalSecret.
It looks like we moved a ClusterExternalSecret
(`argo-notifications-secret`) into an ArgoCD-managed chart
(`argo-services`) but never gave Argo CD permission to create it. This
only showed up when deleting everything out of the staging cluster and
checking whether Argo CD could re-create it all.

Alternatively, we could move this ClusterExternalSecret to the
`argo-bootstrap` chart if we wanted to keep the permissions the same. I
don't think this permissions change represents an actual widening of
access in practice though, as far as I can tell (but happy to be told
otherwise).