Refactor monitoring-config chart.
We don't want to have to copy/paste boilerplate Helm templates just to
add a dashboard file or a new set of alert rules.

The Helm templates monitoring-config/templates are now simply for
collating the static json files under dashboards/ and rules/ into the
k8s objects expected by Prometheus Operator.

Using Prometheus Operator for this stuff probably isn't giving us any
value in the first place, so it might make sense to switch everything to
plain ConfigMaps, but that'd be a bigger change which could use a bit
more thought.