Use same command for Smokey in post sync workflow
This refactors the container command in the post sync workflow to match
the command used to run smokey in the K8s cronjob. We don't need to only
select replatforming specific tests as, the platform should be fully
functional. This also switches the syntax to using multiple --tags flags
instead of "ands" as it harder to read.

For some reason the "configMapKeyRef" configuration to retrieve the
env var value from a Config Map isn't working. Cannot seem to fix, so
using Helm values for now - however would be better to use th ConfigMap
to minimise duplication of value configuration.