Add preStop hook wait for ALB to drain connections
We have a race condition during a rollout between a pod being terminated
and the ALB deregistering it. Currently these process happen
independently and frequently the pods terminate before ALB
deregistration which cause traffic to be sent to a terminating pod. This
causes a few 502/504s to be served until deregistration is complete.
Until we have a better fix, adding a delay until pod termination makes
this senario less likely. However, it would be ideal if we had proper
fix whereby pods are only terminated after it's known that they've be
successfully deregistered from the ALB.