Disable automatic deploys for apps by default
This reverts a fix introduced in #453c36b. If an app was missing the
automatic_deploys_enabled value in the image tag file it would default
to true. This changes the default to false and now requires that value
to be explicitly set. This should minimise implict behaviour and be less
risky for apps that shouldn't be automatically deployed.