Fix wrong security group on account-api ingress LB.
This was open to Internet traffic on the old EC2 setup but accidentally
firewalled off because of a copy-pasto in #949 (by me). Presumably
didn't spot the problem because I happened to be connected to the office
VPN when I tested the fix :(

This opens up traffic to account-api.publishing.service.gov.uk to the
public again, like it was before we switched to Kubernetes.

Verified that the old SG rules allowed this traffic by checking the in
the AWS web console. In prod, rule `sgr-09af9510f487c75ee` was allowing
TCP 443 inbound from `0.0.0.0/0` on `govuk_account_elb_external_access`
(`sg-07787b07d2928de53`).