Add environment variable to Specialist Finder
This commit adds the environment variable
`PUBLISH_PRE_PRODUCTION_FINDERS` to the Specialist Publisher app in
integration, so that finders can be published and tested by end users.

This environment variable was present in the previous hosting
infrastructure, but was not migrated over to the new EKS hosting
platform.

It should only be present in integration. It's been deliberately omitted
from the staging and production environments.