Rightsize resource requests for Release app.
WSS has never gone above 200 MiB and CPU usage is tiny.

Also reduce server replica count to 2 (since this is an internal
dashboard with a very low request rate) and worker count to 1 (since all
the worker does is forward messages to Slack, so if it gets reschedule
then it'll just add a few seconds of delay to an already not-guaranteed
message).

This offsets the extra cost of the recently-added worker deployment.