Permit ingress to asset metadata routes for Asset Manager in integration
We would like to use the integration asset manager to allow development
of new features locally, without needing to replicate the entirety of
Asset Manager's database and assets.

This will permit ingress to the relevant routes and allow us to obtain
that metadata.

The Asset Manager application requires a user to be authenticated to
access this path, so it will only be available to those with a relevant
Signon account or bearer token.