Add chart for Elasticsearch index copy cronjobs.
This replaces the Puppet-managed cronjobs for copying Elasticsearch
indices from production to staging (and staging to integration).

(Neither this nor its predecessor is in involved in production
backups/DR; backups are handled entirely within AWS Managed
Elasticsearch.)

For now, this is its own separate Helm chart. It might become a subchart
of a larger "environment sync" chart later; we'll see what fits.

I've tried to keep this chart close enough to the standard `helm create`
starter chart (minus the templates that we definitely won't use in the
foreseeable future and some that can be trivially back added in if we
want them, like serviceaccount) that it shouldn't be too difficult to
fit it into whatever structure we end up with re what jobs go in what
charts. I.e. it's intended to be composable while avoiding unnecessary
complexity or esotericism.

[Puppet-managed cronjobs]:
https://github.com/alphagov/govuk-puppet/blob/603dc08/hieradata_aws/class/staging/search.yaml#L4