Update hostnames and ordering for draft CSV previews
We are currently routing all requests for `draft-assets` to the live
instance of Frontend first.

These requests should go to the draft Frontend instance instead.

Therefore updating the ordering so draft-frontend has higher priority
and specifying a specific hostname for the draft Frontend instance so we
match only draft requests.